+++ DuDoong-Api/src/main/java/band/gosrock/api/order/controller/CartLineController.java
@@ -1,8 +1,8 @@
 package band.gosrock.api.order.controller;
 
 
-import band.gosrock.api.order.service.CreateOrderLineUseCase;
-import band.gosrock.api.order.service.ReadOrderLineUseCase;
+import band.gosrock.api.order.service.CreateCartLineUseCase;
+import band.gosrock.api.order.service.ReadCartLineUseCase;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -15,12 +15,12 @@ import org.springframework.web.bind.annotation.RestController;
 @SecurityRequirement(name = "access-token")
 @Tag(name = "장바구니 관련 컨트롤러")
 @RestController
-@RequestMapping("/v1/orderLines")
+@RequestMapping("/v1/cartLines")
 @RequiredArgsConstructor
-public class OrderLineController {
+public class CartLineController {
 
-    private final CreateOrderLineUseCase createOrderLineUseCase;
-    private final ReadOrderLineUseCase readOrderLineUseCase;
+    private final CreateCartLineUseCase createOrderLineUseCase;
+    private final ReadCartLineUseCase readOrderLineUseCase;
 
     @Operation(summary = "티켓상품의 갯수와 사용자가 입력한 옵션을 가진 orderLine 을 만듭니다.")
     @PostMapping("/ticketItems/{ticketItem_id}")
