+++ DuDoong-Api/src/main/java/band/gosrock/api/order/service/CreateCartLineUseCase.java
@@ -4,9 +4,7 @@ package band.gosrock.api.order.service;
 import band.gosrock.common.annotation.UseCase;
 
 @UseCase
-public class ReadOrderLineUseCase {
+public class CreateCartLineUseCase {
 
     public void execute() {}
-
-    public void getRecentOrderLine() {}
 }
