+++ DuDoong-Api/src/main/java/band/gosrock/api/order/service/ReadCartLineUseCase.java
@@ -4,7 +4,9 @@ package band.gosrock.api.order.service;
 import band.gosrock.common.annotation.UseCase;
 
 @UseCase
-public class CreateOrderLineUseCase {
+public class ReadCartLineUseCase {
 
     public void execute() {}
+
+    public void getRecentOrderLine() {}
 }
