+++ DuDoong-Domain/src/main/java/band/gosrock/domain/domains/order/domain/OrderId.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class OrderId implements Serializable {
 
-    @Column(name = "id")
+    @Column(name = "order_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long value;
 }
