+++ DuDoong-Domain/src/main/java/band/gosrock/domain/domains/order/domain/OrderLineItemId.java
@@ -1,4 +1,4 @@
-package band.gosrock.domain.domains.order;
+package band.gosrock.domain.domains.order.domain;
 
 
 import java.io.Serializable;
@@ -13,9 +13,9 @@ import lombok.NoArgsConstructor;
 @Getter
 @Embeddable
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class OrderLineId implements Serializable {
+public class OrderLineItemId implements Serializable {
 
-    @Column(name = "id")
+    @Column(name = "order_line_item_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long value;
 }
