+++ DuDoong-Domain/src/main/java/band/gosrock/domain/domains/order/domain/PaymentMethod.java
@@ -1,15 +1,17 @@
 package band.gosrock.domain.domains.order.domain;
 
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+
 @Getter
 @AllArgsConstructor
 public enum PaymentMethod {
-    //간편결제
+    // 간편결제
     EASYPAY("EASYPAY"),
-    //카드결제
+    // 카드결제
     CARD("CARD"),
-    //승인결제
+    // 승인결제
     APPROVAL("APPROVAL");
     private String value;
 }
